change type

+++ utils.py
@@ -91,13 +91,13 @@ def select_index_discriminator(output_discriminator, label):
 def add_zero(tensor,device):
     batch_size = tensor.shape[0]
     new_zero = torch.zeros((batch_size,1)).to(device)
-    tensor = torch.cat([tensor,new_zero],dim =1)
+    tensor = torch.cat([tensor,new_zero],dim =1).double()
     return tensor.to(device)
 
 def create_fake_label(tensor,device):
     batch_size = tensor.shape[0]
     column_size = tensor.shape[1]
-    fake_label = torch.zeros((batch_size,column_size+1))
+    fake_label = torch.zeros((batch_size,column_size+1),dtype = torch.float32)
     fake_label[:,-1] = torch.ones((batch_size,)).to(device)
     return fake_label.to(device)
